fix: Push anchorless URL to history when needed
The goal of these bits of JavaScript are to allow one to click through
the anchor links in the Table of Contents and hit the browser's back key
once to scroll back up to the top of the page and then hit it a second
time to go back to the previous page.

The default, unmodified behavior would mean spamming the back key $n$
times, which is more annoying than helpful. It's a bigger problem on
bloated sites with dynamic content because all the moving elements and
lag invariably resulting in you overshooting your target, but still a
problem nonetheless.

Anyways, this fixes a bug where the anchorless version of the page would
be pushed into the history before the first anchor link click, meaning
it'd take two consecutive back key presses to actually go backwards;
which is a great way to simulate lag.