feat: Convert "Examples" blockquotes into <details> blocks
Over the years I've gotten in the habit of putting examples into
blockquotes like:

> **Example**: Conversion of etoin to shrdlu via stack
>
> 1. Lorem ipsum dolor

So I've added a Pandoc filter that will turn them into <details>
elements for better readability.

This plays well with my Neovim config, which folds these blockquotes
for me, ensuring both the Markdown and output are easier to read.