update: Elab. on Interrupts, User/Kernel Mode, System/Supervisor calls
Also made a miscellenia page that will be a collection-bin of unsorted
information.

Generally I have to do a large-scale reorganization of my notes once I
have a semester's worth of stuff, so I'm going to try to get ahead of
the game by taking the obviously out-of-place information and stuffing
it into this page to save future me some work.