tweak: General Education -> Liberal Arts
The other alternative was "Humanities & Social Sciences", but that
would've been super long.

This is essentially the catch-all bucket for General Education courses
that I managed to take any sort of meaningful notes in.

In terms of future additions, I have some old notes from a history
course that would take forever to clean up and add, and that's about it.