refactor: Combine Monte-Carlo Algorithms into one note
This note is really only Monte-Carlo Integration, split up into
two semi-distinct methods (explicit and implicit shapes)

This distinction was NOT made in class really beyond a header saying
"Multi-Variable Integration", but that's not really a fair split because
if the Multi-Variable Integration is on a shape that isn't closed, it
isn't actually anything new.

Really, I think both of these methods are the same thing but I can't
explain it even though it seems simple.